+++ store/snapshots/manager.go
@@ -551,3 +551,9 @@ func (m *Manager) snapshot(height int64) {
 		m.logger.Debug("pruned state snapshots", "pruned", pruned)
 	}
 }
+
+// Close the snapshot database.
+func (m *Manager) Close() error {
+	m.logger.Info("snapshotManager Close Database")
+	return m.store.db.Close()
+}